Fix Some Cast Warnings  (#40)
* Use size_t for several container sizes

* theta is double

* larger int

* signed-unsigned

* More signed-unsigned

+++ hdi/dimensionality_reduction/gradient_descent_tsne_texture.h
@@ -85,7 +85,7 @@ namespace hdi {
       void getEmbeddingPosition(scalar_vector_type& embedding_position, data_handle_type handle)const;
 
       //! Get the number of data points
-      unsigned int getNumberOfDataPoints() { return _P.size(); }
+      size_t getNumberOfDataPoints() { return _P.size(); }
       //! Get P
       const sparse_scalar_matrix_type& getDistributionP()const { return _P; }
       //! Get Q
